Split args-flow tests into args_tests.rs (file size discipline)
src/app/runtime/tests/mod.rs grew to 1115 lines after Phase 3 added the
three args-shape pin tests. Move custom_args_shapes, the multi-Runtime
parallelism test, and the init-called-once test into a new sibling file
under the existing tests submodule.

No behavior change.
